requirements.txt
Django==4.2.4
djangorestframework==3.14.0
Markdown==3.4.4
gunicorn==21.2.0
kafka-python==2.0.2
djangorestframework-simplejwt==5.3.0
djangorestframework-simplejwt[crypto]
python-dotenv==1.0.0
boto3==1.28.31
django-storages==1.13.2
Pillow==10.0.0
psycopg2-binary==2.9.7
drf-spectacular==0.26.4
django-jsonform==2.19.0
django-prometheus==2.3.1
django-jalali-date==1.0.2
PyJWT==2.8.0
PyJWT[crypto]
django-redis==5.3.0
django-nested-admin==4.0.2
sentry-sdk==1.29.2
django-safedelete==1.3.2
django-cors-headers==4.2.0
django-cacheops==7.0.1
pycryptodome==3.18.0
requests==2.31.0
requests[socks]
